Cleans code
