+++ docker-compose.yml
@@ -23,14 +23,14 @@ services:
     restart: always
     ports:
     - 9000:9000
-#  crypto-analyzer-service:
-#    build: ./crypto-analyzer
-#    depends_on:
-#      - realtime-db
-#    container_name: crypto_analyzer
-#    restart: always
-#    environment:
-#      RETHINK_DB_HOST: ${RETHINK_DB_HOST}
-#      COINBASE_ADAPTER_BASIC_AUTH_USER: ${COINBASE_ADAPTER_BASIC_AUTH_USER}
-#      COINBASE_ADAPTER_BASIC_AUTH_PWD: ${COINBASE_ADAPTER_BASIC_AUTH_PWD}
-#      CB_HOST: ${CB_HOST}
+  crypto-analyzer-service:
+    build: ./crypto-analyzer
+    depends_on:
+      - realtime-db
+    container_name: crypto_analyzer
+    restart: always
+    environment:
+      RETHINK_DB_HOST: ${RETHINK_DB_HOST}
+      COINBASE_ADAPTER_BASIC_AUTH_USER: ${COINBASE_ADAPTER_BASIC_AUTH_USER}
+      COINBASE_ADAPTER_BASIC_AUTH_PWD: ${COINBASE_ADAPTER_BASIC_AUTH_PWD}
+      CB_HOST: ${CB_HOST}
